Don't pass additional requirements separately to the resolver
I tried to upgrade a manageiq (https://github.com/ManageIQ/manageiq)
lockfile generated with older versions of bundler to use the new format
with separated GEM sections, using `bundle _2.2.19_ lock --update` and I
got the following error:

```
Could not find gem 'mime-types (~> 3.0)' in rubygems repository https://rubygems.manageiq.org/ or installed locally.
The source contains the following versions of 'mime-types': 3.0.0
```

This error is absurd.

The explanation is that:

* The previous lockfile had mime-types locked to 3.3.1 from
  rubygems.org, and since bundler 2.2.16, bundler refuses to downgrade
  any top-level dependencies by adding additional ">= <locked_version>"
  constraints to those dependencies. So the resolver is using a
  "mime-types >= 3.3.1" constraint.
* The new secure source resolution logic in bundler 2.2.18 or newer says
  that mime-types needs to be actually fetched from the private
  rubygems.manageiq.org source, but that private source only includes
  mime-types 3.0.0.
* Before resolution starts, bundler checks that all top level
  requirements can actually be found. This list of requirements includes
  top-level dependencies in the Gemfile, but not the additional
  requirements mentioned in the first bullet point. However, when
  actually looking for each depenency, the resolver does consider any
  additional requirements, so it fails to find any matching version in
  the rubygems.manageiq.org source.
* So the reason for not finding mime-types is actually the `mime-types
  >= 3.3.1` additional requirement but the error message does not mention
  it, making it look absurd.

This change merges additional ">=" requirements into the original
Gemfile dependencies, so that the error message now reads like this:

Which is still a bit cryptic (it's not straighforward that the >= 3.3.1
requirement comes from the lockfile), but much better that an absurd
error.